Move Application Header to Shared Components
