+++ alcs-frontend/src/app/shared/application-header/application-header.component.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/colors';
+@import '../../../styles/colors';
 
 
 .application-type {
