+++ alcs-frontend/src/app/shared/shared.module.ts
@@ -27,6 +27,7 @@ import { MtxButtonModule } from '@ng-matero/extensions/button';
 import { DatetimeAdapter } from '@ng-matero/extensions/core';
 import { NgOptionHighlightModule } from '@ng-select/ng-option-highlight';
 import { NgSelectModule } from '@ng-select/ng-select';
+import { ApplicationHeaderComponent } from './application-header/application-header.component';
 import { AvatarCircleComponent } from './avatar-circle/avatar-circle.component';
 import { FavoriteButtonComponent } from './favorite-button/favorite-button.component';
 import { InlineEditComponent } from './inline-edit/inline-edit.component';
@@ -45,6 +46,7 @@ import { MatheoDatepickerFormatter } from './utils/matheo-datepicker-formatter';
     StartOfDayPipe,
     MeetingOverviewComponent,
     InlineEditComponent,
+    ApplicationHeaderComponent,
   ],
   imports: [
     CommonModule,
@@ -55,6 +57,7 @@ import { MatheoDatepickerFormatter } from './utils/matheo-datepicker-formatter';
     FormsModule,
     MatTooltipModule,
     MatCardModule,
+    MatMenuModule,
   ],
   exports: [
     CommonModule,
@@ -73,15 +76,12 @@ import { MatheoDatepickerFormatter } from './utils/matheo-datepicker-formatter';
     MatSnackBarModule,
     MatToolbarModule,
     MatMenuModule,
-    FavoriteButtonComponent,
-    AvatarCircleComponent,
     MatTableModule,
     MatNativeDateModule,
     MatDatepickerModule,
     MatProgressSpinnerModule,
     MatRadioModule,
     MomentPipe,
-    MeetingOverviewComponent,
     MatExpansionModule,
     MtxButtonModule,
     StartOfDayPipe,
@@ -89,6 +89,10 @@ import { MatheoDatepickerFormatter } from './utils/matheo-datepicker-formatter';
     InlineEditComponent,
     MatAutocompleteModule,
     MatButtonToggleModule,
+    ApplicationHeaderComponent,
+    MeetingOverviewComponent,
+    FavoriteButtonComponent,
+    AvatarCircleComponent,
   ],
 })
 export class SharedModule {
